Rename quest to make it specific to grit bins
